requirements.txt
requests==2.9.1
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
django-ckeditor==5.0.3
cloudinary==1.4.0
djangorestframework
markdown
django-filter
django-pagseguro2
coverage==3.7.1
django-extensions
django-password-reset
djangorestframework-jwt
flake8
pep8
pep8-naming
pyflakes
pyjwt
pylint
pylint-django
selenium
unittest-xml-reporting
